tests/wpt/tests/css/selectors/media/sound-state.html

<!DOCTYPE html>
<title>Sound State: the :muted and :volume-locked pseudo-classes</title>
<link
  rel="help"
  href="https://drafts.csswg.org/selectors/#sound-state"
/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<body>
  <video width="300" height="300" loop></video>
  <script type="module">
    // Unfortunately, we can't test the volume-locked state because it's not
    // possible to lock the volume of a video element with JS.

    test((t) => {
      for (const pseudo of [":muted", ":volume-locked"]) {
        try {
          document.querySelector(`.not-a-thing${pseudo}`);
        } catch (e) {
          assert_unreached(`${pseudo} is not supported`);
        }
      }
    }, "Test :pseudo-class syntax is supported without throwing a SyntaxError");

    promise_test(async (t) => {
      assert_implements(CSS.supports("selector(:muted)"), ":muted is not supported");
      assert_equals(
        document.querySelector("video:muted"),
        null,
        "must know :muted"
      );
      const video = document.querySelector("video");
      await new Promise((r) => {
        video.addEventListener("canplay", r, { once: true });
        video.src = getVideoURI("/media/counting");
      });
      video.muted = false;
      assert_false(video.muted, "video is unmuted");
      assert_equals(document.querySelector("video:muted"), null);
      assert_equals(document.querySelector("video:not(:muted)"), video);
      video.muted = true;
      assert_equals(document.querySelector("video:muted"), video);
      assert_equals(document.querySelector("video:not(:muted)"), null);
    }, "Test :muted pseudo-class");

    test((t) => {
      assert_implements(CSS.supports("selector(:muted)"), ":muted is not supported");

      const video = document.createElement("video");
      document.body.appendChild(video);
      t.add_cleanup(() => video.remove());

      assert_false(video.muted);
      assert_false(video.matches(":muted"));

      video.setAttribute("muted", "");
      assert_true(video.muted);
      assert_true(video.matches(":muted"));

      video.removeAttribute("muted");
      assert_false(video.muted);
      assert_false(video.matches(":muted"));

      video.setAttribute("muted", "");
      assert_true(video.muted);
      video.muted = false;
      assert_false(video.muted);
      assert_false(video.matches(":muted"));
    }, "Test :muted matches .muted for content attribute default");
  </script>
</body>
